backend/requirements.txt
appdirs==1.4.4
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
cycler==0.10.0
defusedxml==0.6.0
distlib==0.3.1
Django==2.1.15
django-allauth==0.42.0
django-cors-headers==3.4.0
django-email-utils==0.1.1
django-extensions==3.0.3
django-extra-fields==2.0.5
django-rest-auth==0.9.5
django-rest-email-auth==2.1.0
django-rest-passwordreset==1.1.0
django-webpack-loader==0.7.0
djangorestframework==3.11.0
filelock==3.0.12
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
kiwisolver==1.1.0
matplotlib==3.0.3
mysql==0.0.2
mysql-connector-python==8.0.21
mysqlclient==1.3.12
numpy==1.18.5
oauthlib==3.1.0
Pillow==7.2.0
pprint==0.1
protobuf==3.12.2
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
urllib3==1.25.9
virtualenv==20.0.27
zipp==1.2.0
